community/views.py
from rest_framework import viewsets

from community.models import Community
from community.serializers import CommunitySerializer


class CommunityViewSet(viewsets.ModelViewSet):
    queryset = Community.objects.all()
    serializer_class = CommunitySerializer
